- refactor

+++ controllers/TwilioFaxClient.php
@@ -30,8 +30,8 @@ class TwilioFaxClient extends AppDispatch
     public function __construct()
     {
         $this->crypto = new CryptoGen();
-        $this->baseDir = $GLOBALS['OE_SITE_DIR'] . "/messageStore";
-        $this->uriDir = $GLOBALS['OE_SITE_WEBROOT'] . "/messageStore";
+        $this->baseDir = $GLOBALS['OE_SITE_DIR'] . "/message_store";
+        $this->uriDir = $GLOBALS['OE_SITE_WEBROOT'] . "/message_store";
         $this->cacheDir = $GLOBALS['OE_SITE_DIR'] . '/documents/logs_and_misc/_cache';
         $this->credentials = $this->getCredentials();
         parent::__construct();
@@ -201,12 +201,6 @@ class TwilioFaxClient extends AppDispatch
             return json_encode($ee);
         };
         try {
-            $messageStoreDir = $this->baseDir;
-
-            //Create the Directory
-            if (!file_exists($messageStoreDir)) {
-                mkdir($messageStoreDir, 0777, true);
-            }
             // dateFrom and dateTo paramteres
             $timeFrom = 'T00:00:01Z';
             $timeTo = 'T23:59:59Z';
@@ -238,7 +232,7 @@ class TwilioFaxClient extends AppDispatch
                     $d2 = new DateTime();
                     $dif = $d1->diff($d2);
                     $interval = ($dif->d * 24) + $dif->h;
-                    if ($interval >= 12 && ($status != 'delivered' || $status != 'received')) {
+                    if ($interval >= 12 && ($status != 'delivered' && $status != 'received')) {
                         $f = $twilio->fax->v1->faxes($id)->delete();
                     } elseif ($interval >= 48) {
                         $f = $twilio->fax->v1->faxes($id)->delete();
